_mrrc::wrappers

Struct PyHoldingsRecord

pub struct PyHoldingsRecord {
    pub inner: HoldingsRecord,
}

Python wrapper for a MARC Holdings Record (Type x/y/v/u)

Holdings records are used to maintain inventory and location information. They use the same ISO 2709 binary format as bibliographic records but organize fields by functional role (locations, enumeration, notes, etc.).

Fields
inner: HoldingsRecord
Implementations

impl PyHoldingsRecord

pub fn leader(&self) -> PyLeader

Get the leader

pub fn record_type(&self) -> String

Get record type (single character: x, y, v, or u)

pub fn locations(&self) -> Vec<PyField>

Get all location fields (852)

pub fn captions_basic(&self) -> Vec<PyField>

Get all basic captions (853)

pub fn captions_supplements(&self) -> Vec<PyField>

Get all supplement captions (854)

pub fn captions_indexes(&self) -> Vec<PyField>

Get all index captions (855)

pub fn enumeration_basic(&self) -> Vec<PyField>

Get all basic enumeration (863)

pub fn enumeration_supplements(&self) -> Vec<PyField>

Get all supplement enumeration (864)

pub fn enumeration_indexes(&self) -> Vec<PyField>

Get all index enumeration (865)

pub fn textual_holdings_basic(&self) -> Vec<PyField>

Get all basic textual holdings (866)

pub fn textual_holdings_supplements(&self) -> Vec<PyField>

Get all supplement textual holdings (867)

pub fn textual_holdings_indexes(&self) -> Vec<PyField>

Get all index textual holdings (868)

pub fn get_fields(&self, tag: &str) -> Option<Vec<PyField>>

Get fields by tag

pub fn get_control_field(&self, tag: &str) -> Option<String>

Get control field by tag

pub fn to_json(&self) -> PyResult<String>

Convert to JSON string

fn __repr__(&self) -> String

fn __str__(&self) -> String