_mrrc::wrappers

Struct PyRecord

pub struct PyRecord {
    pub inner: Record,
}

Python wrapper for a Record

A MARC bibliographic record is the fundamental unit of MARC data. It consists of a leader (24-byte header), control fields (000-009), and data fields (010 and above).

Examples

import mrrc
leader = mrrc.Leader()
leader.record_type = 'a'
leader.bibliographic_level = 'm'

record = mrrc.Record(leader)
record.add_control_field('001', '12345')

field = mrrc.Field('245', '1', '0')
field.add_subfield('a', 'Example Title')
record.add_field(field)

print(f"Title: {record.title()}")

Fields
inner: Record
Implementations

impl PyRecord

pub fn new(leader: &PyLeader) -> Self

Create a new Record with a given Leader

pub fn leader(&self) -> PyLeader

Get the leader

pub fn add_control_field(&mut self, tag: &str, value: &str) -> PyResult<()>

Add a control field (000-009)

pub fn control_field(&self, tag: &str) -> Option<String>

Get a control field value

pub fn add_field(&mut self, field: &PyField)

Add a data field

pub fn fields_by_tag(&self, tag: &str) -> Vec<PyField>

Get fields by tag

pub fn get_field(&self, tag: &str) -> Option<PyField>

Get the first field with a given tag (pymarc compatibility)

pub fn get_fields(&self, tag: &str) -> Vec<PyField>

Get all fields with a given tag (pymarc compatibility)

pub fn fields(&self) -> Vec<PyField>

Get all fields

pub fn control_fields(&self) -> Vec<(String, String)>

Get all control fields as a dict-like structure

pub fn remove_field(&mut self, tag: &str) -> Vec<PyField>

Remove all fields with a given tag

Returns the removed fields.

pub fn title(&self) -> Option<String>

Get title from 245 field (first subfield $a)

pub fn author(&self) -> Option<String>

Get author from 100/110/111 field

pub fn isbn(&self) -> Option<String>

Get ISBN from 020 field

pub fn subjects(&self) -> Vec<String>

Get all subject headings from 650 field

pub fn location(&self) -> Vec<String>

Get all location fields (852)

pub fn notes(&self) -> Vec<String>

Get all notes from 5xx fields

pub fn publisher(&self) -> Option<String>

Get publisher from 260 field

pub fn uniform_title(&self) -> Option<String>

Get uniform title from 130 field

pub fn sudoc(&self) -> Option<String>

Get SuDoc (government document classification) from 086 field

pub fn issn_title(&self) -> Option<String>

Get ISSN title from 222 field

pub fn issnl(&self) -> Option<String>

Get ISSN-L from 024 field

pub fn pubyear(&self) -> Option<u32>

Get publication year (alias for publication_year)

pub fn issn(&self) -> Option<String>

Get ISSN from 022 field

pub fn series(&self) -> Option<String>

Get series from 490 field

pub fn physical_description(&self) -> Option<String>

Get physical description from 300 field

pub fn is_book(&self) -> bool

Check if this record is a book (record type ‘a’ + bibliographic level ‘m’)

pub fn is_serial(&self) -> bool

Check if this record is a serial (bibliographic level ‘s’)

pub fn is_music(&self) -> bool

Check if this record is music (record type ‘c’ or ‘d’)

pub fn is_audiovisual(&self) -> bool

Check if this record is audiovisual material (record type ‘g’)

pub fn to_json(&self) -> PyResult<String>

Convert record to JSON string

Example

json_str = record.to_json()

pub fn to_xml(&self) -> PyResult<String>

Convert record to XML string

Example

xml_str = record.to_xml()

pub fn to_marcjson(&self) -> PyResult<String>

Convert record to MARCJSON format

Example

marcjson_str = record.to_marcjson()

pub fn to_dublin_core(&self) -> PyResult<HashMap<String, Vec<String>>>

Convert record to Dublin Core metadata

Returns

A dictionary mapping Dublin Core field names to lists of values

Example

dc = record.to_dublin_core()
print(dc['title'])

pub fn to_mods(&self) -> PyResult<String>

Convert record to MODS XML format

MODS (Metadata Object Description Schema) is a more detailed XML representation than Dublin Core.

Example

mods_xml = record.to_mods()

pub fn to_marc21(&self) -> PyResult<Vec<u8>>

Convert record to MARC21 binary format (ISO 2709)

Returns the record as bytes in MARC21 format following the ISO 2709 standard. This is the standard MARC binary interchange format used by libraries worldwide.

Returns

Bytes object containing the MARC21-encoded record

Example

marc_bytes = record.to_marc21()
with open('record.mrc', 'wb') as f:
    f.write(marc_bytes)

fn __repr__(&self) -> String

fn __str__(&self) -> String

fn __eq__(&self, other: &PyRecord) -> bool

impl PyRecord

unsafe fn __pymethod___eq____( py: Python<'_>, _raw_slf: *mut PyObject, arg0: *mut PyObject, ) -> PyResult<*mut PyObject>

impl PyRecord

unsafe extern "C" fn __pymethod___richcmp____( slf: *mut PyObject, other: *mut PyObject, op: c_int, ) -> *mut PyObject